+++ src/test/java/uk/gov/cslearning/catalogue/api/CourseControllerTest.java
@@ -3,7 +3,6 @@ package uk.gov.cslearning.catalogue.api;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.google.common.collect.ImmutableMap;
 import org.glassfish.jersey.servlet.WebConfig;
 import org.junit.Before;
 import org.junit.Test;
@@ -25,26 +24,22 @@ import uk.gov.cslearning.catalogue.domain.CivilServant.OrganisationalUnit;
 import uk.gov.cslearning.catalogue.domain.Course;
 import uk.gov.cslearning.catalogue.domain.Status;
 import uk.gov.cslearning.catalogue.domain.Visibility;
-import uk.gov.cslearning.catalogue.domain.module.*;
 import uk.gov.cslearning.catalogue.repository.elastic.CourseRepository;
-import uk.gov.cslearning.catalogue.service.*;
+import uk.gov.cslearning.catalogue.service.CourseService;
+import uk.gov.cslearning.catalogue.service.RegistryService;
 
-import java.net.URL;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.*;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Optional;
 
-import static org.hamcrest.Matchers.*;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.when;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-import static uk.gov.cslearning.catalogue.exception.ResourceNotFoundException.resourceNotFoundException;
 
 
 @RunWith(SpringJUnit4ClassRunner.class)
@@ -62,15 +57,6 @@ public class CourseControllerTest {
 
     @MockBean
     private CourseService courseService;
-
-    @MockBean
-    private ModuleService moduleService;
-
-    @MockBean
-    private EventService eventService;
-
-    @MockBean
-    private AudienceService audienceService;
 
     @MockBean
     private RegistryService registryService;
@@ -223,351 +209,6 @@ public class CourseControllerTest {
                                 .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.title", equalTo("title")));
-    }
-
-    @Test
-    public void shouldCreateModule() throws Exception {
-        String moduleId = "module-id";
-        Module module = mock(LinkModule.class);
-        when(module.getId()).thenReturn(moduleId);
-
-        String courseId = UUID.randomUUID().toString();
-        String json = objectMapper.writeValueAsString(ImmutableMap.of("type", "link", "location", "http://localhost"));
-
-        when(moduleService.save(eq(courseId), any(Module.class))).thenReturn(module);
-
-        mockMvc.perform(
-                        post(String.format("/courses/%s/modules/", courseId)).with(csrf())
-                                .content(json)
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isCreated())
-                .andExpect(header().string("Location", String.format("http://localhost/courses/%s/modules/%s", courseId, moduleId)));
-    }
-
-    @Test
-    public void shouldFindModule() throws Exception {
-        String courseId = "course-id";
-        String moduleId = "module-id";
-        String url = "http://example.org";
-
-        Module module = new LinkModule(new URL(url));
-
-        when(moduleService.find(courseId, moduleId)).thenReturn(Optional.of(module));
-
-        mockMvc.perform(
-                        get(String.format("/courses/%s/modules/%s", courseId, moduleId)).with(csrf())
-                                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.url", equalTo(url)));
-    }
-
-    @Test
-    public void shouldReturnNotFoundIfModuleNotFound() throws Exception {
-        String courseId = "course-id";
-        String moduleId = "module-id";
-
-        when(courseRepository.existsById(courseId)).thenReturn(true);
-        when(moduleService.find(courseId, moduleId)).thenReturn(Optional.empty());
-
-        mockMvc.perform(
-                        get(String.format("/courses/%s/modules/%s", courseId, moduleId)).with(csrf())
-                                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    public void shouldDeleteModule() throws Exception {
-        String courseId = "course-id";
-        String moduleId = "module-id";
-
-        doNothing().when(moduleService).deleteModule(courseId, moduleId);
-
-        mockMvc.perform(
-                        delete(String.format("/courses/%s/modules/%s", courseId, moduleId)).with(csrf())
-                                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNoContent());
-    }
-
-    @Test
-    public void shouldUpdateModule() throws Exception {
-        String courseId = "course-id";
-        String moduleId = "id-123";
-        String title = "old-title";
-        String updatedTitle = "updated-title";
-        String url = "https://www.example.org";
-
-        Course course = new Course();
-        Course updatedCourse = new Course();
-
-        Module module = new LinkModule(new URL(url));
-        module.setId(moduleId);
-        module.setTitle(title);
-
-        Module updatedModule = new LinkModule(new URL(url));
-        updatedModule.setId(moduleId);
-        updatedModule.setTitle(updatedTitle);
-
-        List<Module> modules = new ArrayList<>();
-        modules.add(module);
-        course.setModules(modules);
-
-        List<Module> updatedModules = new ArrayList<>();
-        updatedModules.add(updatedModule);
-        updatedCourse.setModules(updatedModules);
-
-        when(moduleService.find(any(), any())).thenReturn(Optional.of(module));
-        when(moduleService.updateModule(courseId, updatedModule)).thenReturn(updatedCourse);
-
-        mockMvc.perform(
-                        put(String.format("/courses/%s/modules/%s", courseId, module.getId())).with(csrf())
-                                .content(objectMapper.writeValueAsString(updatedModule))
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNoContent());
-
-        assertThat(updatedCourse.getModules().isEmpty(), is(false));
-        assertThat(updatedCourse.getModules().size(), is(1));
-        assertThat(updatedCourse.getModules().get(0).getTitle(), is(updatedTitle));
-    }
-
-
-    @Test
-    public void shouldAddEventToModule() throws Exception {
-        Event event = new Event();
-        event.setJoiningInstructions("");
-        event.setDateRanges(new ArrayList<>());
-        event.setVenue(new Venue("location"));
-
-        String courseId = "course-id";
-        String moduleId = "module-id";
-
-        when(courseRepository.existsById(courseId)).thenReturn(true);
-
-        when(eventService.save(eq(courseId), eq(moduleId), any(Event.class))).thenReturn(event);
-
-        mockMvc.perform(
-                        post(String.format("/courses/%s/modules/%s/events", courseId, moduleId)).with(csrf())
-                                .content(objectMapper.writeValueAsString(event))
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isCreated());
-    }
-
-    @Test
-    public void shouldReturnEvent() throws Exception {
-        String courseId = "course-id";
-        String moduleId = "module-id";
-        String eventId = "event-id";
-
-        Event event = new Event();
-
-        when(courseRepository.existsById(courseId)).thenReturn(true);
-        when(eventService.find(courseId, moduleId, eventId)).thenReturn(Optional.of(event));
-
-        mockMvc.perform(
-                        get(String.format("/courses/%s/modules/%s/events/%s", courseId, moduleId, eventId)).with(csrf())
-                                .content(objectMapper.writeValueAsString(event))
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk());
-    }
-
-    @Test
-    public void shouldReturnNotFoundIfEventNotFound() throws Exception {
-        String courseId = "course-id";
-        String moduleId = "module-id";
-        String eventId = "event-id";
-
-        when(eventService.find(courseId, moduleId, eventId)).thenReturn(Optional.empty());
-
-        mockMvc.perform(
-                        get(String.format("/courses/%s/modules/%s/events/%s", courseId, moduleId, eventId)).with(csrf())
-                                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    public void shouldUpdateEvent() throws Exception {
-
-        LocalDate date = LocalDate.now();
-        LocalTime start = LocalTime.NOON;
-        LocalTime end = LocalTime.MIDNIGHT;
-
-        DateRange dateRange = new DateRange();
-        dateRange.setDate(date);
-        dateRange.setStartTime(start);
-        dateRange.setEndTime(end);
-
-        List<DateRange> dateRanges = Collections.singletonList(dateRange);
-        Venue venue = new Venue();
-        venue.setLocation("venue-location");
-        venue.setAddress("venue-address");
-        venue.setCapacity(10);
-        venue.setMinCapacity(5);
-
-        Course course = new Course();
-
-        Event oldEvent = new Event();
-        Event newEvent = new Event();
-
-        newEvent.setJoiningInstructions("new");
-        newEvent.setDateRanges(dateRanges);
-        newEvent.setVenue(venue);
-        oldEvent.setJoiningInstructions("old");
-
-        FaceToFaceModule module = new FaceToFaceModule("product-code");
-
-        HashSet<Event> events = new HashSet<>();
-        events.add(oldEvent);
-        module.setEvents(events);
-
-        List<Module> modules = new ArrayList<>();
-        modules.add(module);
-        course.setModules(modules);
-
-        when(courseRepository.existsById(course.getId())).thenReturn(true);
-
-        Optional<Course> result = Optional.of(course);
-        when(courseRepository.findById(course.getId())).thenReturn(result);
-
-        when(courseRepository.save(course)).thenReturn(course);
-
-        mockMvc.perform(
-                        put(String.format("/courses/%s/modules/%s/events/%s", course.getId(), module.getId(), oldEvent.getId())).with(csrf())
-                                .content(objectMapper.writeValueAsString(newEvent))
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .accept(MediaType.APPLICATION_JSON))
-                .andDo(print())
-                .andExpect(status().isOk());
-
-        Event savedEvent = module.getEvents().stream().filter(e -> e.getId().equals(oldEvent.getId())).findFirst().get();
-
-        assert (module.getEvents().size() == 1);
-        assertEquals(savedEvent.getId(), oldEvent.getId());
-        assertEquals("new", savedEvent.getJoiningInstructions());
-        assertEquals(venue, savedEvent.getVenue());
-        assertEquals(dateRange, savedEvent.getDateRanges().get(0));
-    }
-
-    @Test
-    public void shouldDeleteEvent() throws Exception {
-        Course course = new Course();
-        FaceToFaceModule module = new FaceToFaceModule("product-code");
-        Event event = new Event();
-
-        HashSet<Event> events = new HashSet<>();
-        events.add(event);
-        module.setEvents(events);
-
-        List<Module> modules = new ArrayList<>();
-        modules.add(module);
-        course.setModules(modules);
-
-        when(courseRepository.existsById(course.getId())).thenReturn(true);
-
-        Optional<Course> result = Optional.of(course);
-        when(courseRepository.findById(course.getId())).thenReturn(result);
-
-        when(courseRepository.save(course)).thenReturn(course);
-
-        mockMvc.perform(
-                        delete(String.format("/courses/%s/modules/%s/events/%s", course.getId(), module.getId(), event.getId())).with(csrf()))
-                .andExpect(status().isNoContent());
-
-        assert (module.getEvents().isEmpty());
-    }
-
-    @Test
-    public void shouldCreateAudience() throws Exception {
-        String audienceId = "audience-id";
-        Audience audience = mock(Audience.class);
-        when(audience.getId()).thenReturn(audienceId);
-
-        String courseId = UUID.randomUUID().toString();
-
-        when(audienceService.isPermitted(any(), any())).thenReturn(true);
-        when(audienceService.setDefaults(any(), any())).thenReturn(audience);
-        when(audienceService.save(any(), any())).thenReturn(new Course());
-
-        mockMvc.perform(
-                        post(String.format("/courses/%s/audiences/", courseId)).with(csrf())
-                                .content(objectMapper.writeValueAsString(ImmutableMap.of("id", audienceId, "name", "Audience name")))
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isCreated())
-                .andExpect(header().string("Location", String.format("http://localhost/courses/%s/audiences/%s", courseId, audienceId)));
-    }
-
-    @Test
-    public void shouldFindAudience() throws Exception {
-        String courseId = "course-id";
-        String audienceId = "audience-id";
-
-        Audience audience = new Audience();
-        audience.setId(audienceId);
-
-        when(courseRepository.existsById(courseId)).thenReturn(true);
-        when(audienceService.find(courseId, audienceId)).thenReturn(Optional.of(audience));
-
-        mockMvc.perform(
-                        get(String.format("/courses/%s/audiences/%s", courseId, audienceId)).with(csrf())
-                                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.id", equalTo(audienceId)));
-    }
-
-    @Test
-    public void shouldReturnNotFoundIfAudienceNotFound() throws Exception {
-        String courseId = "course-id";
-        String audienceId = "audience-id";
-
-        when(courseRepository.existsById(courseId)).thenReturn(true);
-        when(audienceService.find(courseId, audienceId)).thenReturn(Optional.empty());
-
-        mockMvc.perform(
-                        get(String.format("/courses/%s/audiences/%s", courseId, audienceId)).with(csrf())
-                                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    public void shouldReturnNotFoundIfCourseNotFoundWhenFindingAudience() throws Exception {
-        String courseId = "course-id";
-        String audienceId = "audience-id";
-
-        when(audienceService.isPermitted(any(), any())).thenReturn(true);
-        Audience audience = new Audience();
-        when(audienceService.setDefaults(any(), any())).thenReturn(audience);
-
-        doThrow(resourceNotFoundException()).when(audienceService).save(courseId, audience);
-
-        mockMvc.perform(
-                        get(String.format("/courses/%s/audiences/%s", courseId, audienceId)).with(csrf())
-                                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    public void shouldDeleteAudience() throws Exception {
-        Course course = new Course();
-        Audience audience = new Audience();
-        Set<Audience> audiences = new HashSet<>();
-        audiences.add(audience);
-        course.setAudiences(audiences);
-
-        when(courseRepository.findById(course.getId())).thenReturn(Optional.of(course));
-        when(audienceService.find(course, audience.getId())).thenReturn(Optional.of(audience));
-        when(audienceService.isPermitted(any(), any())).thenReturn(true);
-        when(courseRepository.save(course)).thenReturn(course);
-
-        assertThat(course.getAudiences().isEmpty(), is(not(true)));
-
-        mockMvc.perform(
-                        delete(String.format("/courses/%s/audiences/%s", course.getId(), audience.getId())).with(csrf()))
-                .andExpect(status().isNoContent());
-
-        assertThat(course.getAudiences().isEmpty(), is(true));
     }
 
     private Course createCourse() {
